# api.yaml
swagger: '2.0'
info:
  title: Reshigo API
  version: "1.0"
# the domain of the service
host: server.reshigo.ru
...
schemes:
  - http
# will be prefixed to all paths
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: Http Basic Authentication. Works over `HTTP`
paths:
...
  /orders:
    get:
      description: Get available for solving orders.
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: offsetDate
          in: query
          required: false
          description: default value if 2000000000000 (about 2033 year). There are orders with it\'s date less or equal than specified.
          type: integer
          format: int64
        - name: offset
          in: query
          required: false
          description: default offset is 0
          type: integer
          format: int64
        - name: limit
          in: query
          required: false
          description: default limit is 30
          type: integer
          format: int64
        - name: subject
          in: query
          description: |
            subject\'s id, no filter will be applied if not set
          type: integer
          format: int
          required: false
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
  /orders/{id}:
    get:
      description: Get single order from AVAILABLE
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          required: true
          in: path
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Order'
        404:
          description: No such order among AVAILABLE
    put:
      description: Reserve order to solve later. Limit is 2 orders.
      tags:
        - Solver
      security:
        - basicAuth: []
      parameters:
        - name: id
          type: integer
          format: int64
          required: true
          in: path
      responses:
        200:
          description: OK
        403:
          description: The order has already been reserved.
        405:
          description: Solver can't reserve more orders, or this order is only for price suggestions
...
  /orders/{id}/pictures/{picture_id}:
    get:
          description: Get order\'s picture
          tags:
            - Solver
          security:
            - basicAuth: []
          parameters:
            - name: id
              type: integer
              format: int64
              in: path
              required: true
            - name: picture_id
              type: integer
              format: int64
              in: path
              required: true
          responses:
            200:
              schema:
                $ref: '#/definitions/Picture'
              description: OK
            404:
              description: the order or the picture was not found
...
definitions:
...
  User:
    type: object
    properties:
      name:
        type: string
        description: should satisfy regexp [a-zA-Z0-9_/\\\[\]\.\{\}\-]+ and can't consist only from numbers
      email:
        type: string
        description: DEPRECATED
      phone:
        type: string
      fcmToken:
        description: token to use for notifications server
        type: string
      isCustomer:
        type: boolean
      promocode:
        description: sequence of characters to invite friends
        type: string
      funds:
        type: number
        format: double
      reservedFunds:
        type: number
        format: double
      commission:
        type: number
        format: float
      info:
        type: string
        format: binary
        description: utf-8 string
      education:
        type: string
        description: utf-8 string
      degree:
        type: string
        description: utf-8 string, representing class, course or other academic degree
      campaign:
        type: string
        description: source, from where this user came
      iosReview:
        type: string
        description: if true, then user reviewed application in AppStore. can be changed via PUT /user method
      androidReview:
        type: string
        description: if true, then user reviewed application in PlayStore. can be changed via PUT /user method
...
  Picture:
    type: object
    properties:
      id:
        type: integer
        format: int64
      counter:
        type: integer
        format: int64
        description: Field to sort pictures. You had better use unix-timestamp to set this filed.
      img:
        $ref: '#/definitions/PictureData'
  PictureData:
    type: object
    properties:
      data:
        type: string
        format: binary
        description: base64 binary representation of the image
  Order:
    type: object
    description: Every property of theme except \'id\' is ignored.
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
      date:
        type: integer
        format: int64
        description: unix time
      maturityDate:
        type: integer
        format: int64
        description: unix time
      comment:
        type: string
        format: binary
        description: base64 binary representation of string
      theme:
        $ref: '#/definitions/Theme'
      pictures:
        description: This field contains only ids and counters of pictures.
        type: array
        items:
          $ref: '#/definitions/Picture'
      price:
        description: min value is 90
        type: integer
        format: int64
      tasksCnt:
        description: number of tasks
        type: integer
        format: int64
      chat:
        $ref: '#/definitions/Chat'
      user:
        description: Contains customer\'s login and flag isCustomer
        $ref: '#/definitions/User'
      solver:
        description: Contains solver\'s login and flag isCustomer
        $ref: '#/definitions/User'
  Subject:
    type: object
    properties:
      id:
        type: integer
        format: int64
      subject:
        type: string
      subjectRU:
        type: string
        format: binary
  Theme:
    type: object
    properties:
      id:
        type: integer
        format: int64
      theme:
        type: string
      themeRU:
        type: string
        format: binary
      subject:
        $ref: '#/definitions/Subject'
  Chat:
    type: object
    description: All properties of order except id are ignored
    properties:
      id:
        type: integer
        format: int64
      order:
        $ref: '#/definitions/Order'
      status:
        type: string
        description: OPEN, CLOSED
...